feat: 퀀트 멀티팩터 스코어링 모델 도입 및 시스템 고도화
[스코어링 로직 전면 개편]
- 10점 만점(0.5점 단위)의 퀀트 멀티팩터 모델 적용 (Trend, Momentum, Strength, Synergy)
- MACD 지표 추가: 골든크로스(+1.0), 0선 상회(+0.5) 반영
- RSI 강세 구간(50~75) 가점 및 CCI, OBV, ADX 활용 로직 정교화
- 매수 기준 점수(8.0) 및 RSI 상한(65) 등 기본 전략 파라미터 최적화

[시스템 트레이딩 및 분석 기능 강화]
- 전체 종목 분석 및 트레이딩 로그에 SAR, MACD, CCI, OBV 등 상세 지표 출력 추가
- 백테스팅 시뮬레이션 시 상태 분류 로직 동기화 (로그 정확도 향상)
- 종목별 트레이딩 룰 설정 시 소수점 점수 입력 지원

[사용자 경험(UX) 및 문서 개선]
- 시스템 트레이딩 메뉴 네비게이션 구조 변경 (서브 메뉴 실행 후 메인 메뉴로 즉시 복귀)
- README.md 매매 전략, 스코어링 기준, 리스크 관리 정책 상세 기술
- 도움말([H]) 화면에 최신 로직 및 설정값 동적 반영

+++ modules/db_manager.py
@@ -63,9 +63,9 @@ class DBManager:
                     CREATE TABLE IF NOT EXISTS stock_strategies (
                         code TEXT PRIMARY KEY,
                         name TEXT,
-                        buy_score INTEGER,
+                        buy_score REAL,
                         buy_rsi REAL,
-                        sell_score INTEGER,
+                        sell_score REAL,
                         stop_loss REAL,
                         take_profit REAL,
                         take_profit_rsi REAL,
